+++ src/main/java/ufjf/poo/model/pedido/ItemPedido.java
@@ -2,5 +2,5 @@ package ufjf.poo.model.pedido;
 
 import ufjf.poo.model.estoque.Produto;
 
-public record ItemPedido(Produto produto, int quantidade, double subtotal) {
+public record ItemPedido(Produto produto, int quantidade, long subtotal) {
 }
